+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.1.0 (2019-09-19)
+
+New prop `enableColumnWideDrag` to control a column's draggable hotspot - Set to {false} if you experience buggyness when using with react-table column sorting/resizing functionality
+
+Bug fix for calls made to `onDraggedColumnChange` when column ordering has not changed
+
 # 1.0.1 (2019-08-09)
 
 Bug fix for reordering columns not working as expected under certain edge cases